Rename core->config, split config into individual modules
