+++ src/config/external.rs
@@ -1,0 +1,17 @@
+use pyo3::prelude::*;
+use serde::{Deserialize, Serialize};
+
+#[derive(Debug, Serialize, Default, Deserialize, Clone, PartialEq)]
+#[pyclass(get_all, module = "tach.extension")]
+pub struct ExternalDependencyConfig {
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub exclude: Vec<String>,
+    #[serde(default, skip_serializing_if = "Vec::is_empty")]
+    pub rename: Vec<String>,
+}
+
+impl ExternalDependencyConfig {
+    pub fn is_default(&self) -> bool {
+        *self == Self::default()
+    }
+}
