+++ src/interfaces/check.rs
@@ -3,7 +3,7 @@ use std::path::PathBuf;
 use super::compiled::CompiledInterfaces;
 use super::data_types::{TypeCheckCache, TypeCheckResult};
 use super::error::InterfaceError;
-use crate::core::config::{InterfaceConfig, ModuleConfig};
+use crate::config::{InterfaceConfig, ModuleConfig};
 
 pub struct InterfaceChecker {
     interfaces: CompiledInterfaces,
